- refactored crypto-payment integration tests
